Add the idea of "heap variables" in prep for IR and closures
The first version of IR codegen will not be in SSA form and will use
heap variables to represent everything. These will not be only for an
intermediate representation however, in the near future when closures
are properly implemented they will be used for all variables that are
shared between independent functions.