Be much more principled about stack usage in the IR
No more stack handling in the IR. Getting function parameters is its
own dedicated instruction in the IR and also in the VM bytecode. Popping
the arguments off of the stack in the VM bytecode is not allowed, and
every VM instruction that consumes the stack consumes *all* of the stack
(after the function parameters which are immutable).

`Call` and `Method` now have an optional single return. This removes
support for multiple returns within the IR (which is not currently used
anyway) to be replaced in the future by a proper "tuple" IR type. This
keeps the IR in SSA form without stack manipulation (since the return
type of `Call` and `Return` is either `Void` or `Value`). This will all
be necessary in the future for easy inlining.