Rework IR variable system to account for upvalue closing
It is not really possible to use simple analysis passes on the previous
IR to determine when upvalues should be logically "reset" during
function execution.

Instead, we need a new IR instructions to "close" variables when
they logically go out of scope in the original source. We use the new
`ResetHeap` VM instruction to force disconnection of the closed over
variable when it goes out of scope.

We could do this instead at *declaration time* rather than when the
variable goes out of scope, but this is less than optimal because a
variable (outside of a loop anyway) will exist in the parent function's
state for the entire rest of the function.

Additionally, the old allocation system for heap variables was that
every heap variable got its own unique slot, which for large functions
is sub-optimal because it keeps every heap value alive forever.

Instead of doing this, we actually add *two* new IR instructions, one
IR instruction to "open" a variable and one to "close" it, that are
connected to the original variable scope. The "open" instruction is
purely virtual and only used to allocate a heap slot, the "close"
instruction both frees the heap slot and also resets the heap variable
in the bytecode.

In the future there needs to be an optimization pass that removes
variable close instructions that immediately preceed a return in a basic
block.