allow `_` in numbers, don't silent convert large int literals to float
Instead of silently parsing large integers as floats like Lua does,
instead let them *fail* to parse and instead require that the code have
a decimal point or exponent to mark them as a float literal.

It is now possible *at a glance* to see what the type of a numeric
literal is without having to look up the value of `2^63 - 1`.