Add the concept of "this scopes" to the IR, matching VM instructions
"This scopes" are for constructs like `with(obj) {}` where the `self`
and `other` intrinsics change value for a certain scope that can be
broken with `break` etc.

The VM now has instructions for *pushing* new self objects and *popping*
previous self objects off of the built-in stack, and these instructions
correctly modify both `self` and `other` (which is set to previous value
of `self`).

In addition, struct literals now set the `self` value to the struct
being currently constructed, matching GMS2.