Support `Value` as errors internally (support for try / catch / throw)
Removes `anyhow` as a dependency of the VM, makes errors have less
indirection and greatly simplifies error handling externally.

The `vm::Error` grows to 24 bytes (3 words) because it becomes an enum
containing a fat pointer. This may eventually become an issue but it
should be possible to eventually do something like what `anyhow` does
internally combined with pointer tagging to get it down to 8 bytes (1
word).