Only allow concrete error types to be converted to vm::RuntimeError
Still provides interop with anyhow via `RuntimeError::from_boxed`, more
closely matching what anyhow does. This requires double indirection, but
that is also true for anyhow.

This is in preparation for an eventual switch to a *thin*
`vm::RuntimeError`, which requires sized error types.

The change is more verbose but more explicit, and may already have
performance benefits by not require going through Box<Error> to get
to Arc<Error>.