Clean up string handling and printing and support `toString()` methods
This purposefully breaks compatibility with GMS2. Any `toString()`
methods are ONLY called during this *explicit* style of string
conversion, or "printing coercion", which is used by `string` and
`show_debug_message` builtins.

To make everything match as closely as possible with how GMS2 treats
strings, (and to make printing predictable) most other stdlib methods
also use printing coercion for string arguments. Notably the string
length builtins and buffer write methods for strings.

Printing coercion is infallible (other than erroring `toString()`
methods) and can be performed on EVERY value type. The result may be
arbitrary and only suitable for printing for a human to read. Because of
this, printing coercion is NOT used in the VM to implicitly convert the
keys of structs to strings! This only invokes the normal, weaker string
coercion and can error if the value does not make sense as a string and
thus does not make sense as a struct key.