Crop textures based on texture group setting, deduplicate sprite frames
This makes fabricator produce texture atlases that are vastly closer to
what GMS2 will produce.

Fabricator's box packing algorithm can actually be *more agressive*
than the one in GMS2 and can produce smaller atlases, as long as the
`GRANULARITY` constant is turned down to 1. This was verified using GPU
frame captures, since fabricator does not write any atlas files to disk.

What fabricator will *not* do currently that GMS2 will do is produce
smaller atlases even if the group texture size is set to a larger
value... for example, GMS2 will produce 2048x2048 or even a 2048x1024
textures if there is only one texture in a group and all textures will
fit, even if the platform texture group size is set to 4096x4096.

Sprite frames are currently deduplicated by *raw file contents* for
speed and out of an overabundance of correctness, since it is not the
job of the `fabricator` crate to know what the image and texture format
needs to be.