Clean up and simplify calling functions on `Thread`
`Thread` and `Execution` can now both call `Function` in a way
that should produce clean errors in all cases. Rather than have
`ExternVmError`, there is a replacement error type called `CallError`
which is returned from both `Execution::call` and `Thread::run[_with]`.
`Execution::call` *assumes that the caller is using the same `Thread`*
and thus can avoid `CallError` directly containing another `CallError`.
Error types and backtraces are handled intelligently here: whatever
the *innermost* error and backtrace are ends up in the outermost
`CallError`.

This error handling is NOT appropriate for a callback which calls a
function on another thread! This is not a problem because this should
*only* ever happen with coroutines, and when coroutines are implemented,
`Thread::resume` or similar will not return a `CallError`.